# docs/swagger.yaml
basePath: /
definitions:
...
  model.Shipment:
    properties:
      actual_delivery:
        type: string
      cargo_description:
        type: string
      cargo_value:
        type: number
      cargo_weight:
        type: number
      created_at:
        type: string
      created_by:
        type: string
      customer_name:
        type: string
      customer_phone:
        type: string
      destination_address:
        type: string
      destination_lat:
        type: number
      destination_lng:
        type: number
      driver_id:
        type: string
      estimated_delivery:
        type: string
      id:
        type: string
      origin_address:
        type: string
      origin_lat:
        type: number
      origin_lng:
        type: number
      special_instructions:
        type: string
      status:
        type: string
      status_note:
        type: string
      status_reason:
        type: string
      tenant_id:
        type: string
      tracking_number:
        type: string
      updated_at:
        type: string
      vehicle_id:
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.Error'
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
  service.AssignDriverInput:
    properties:
      driver_id:
        type: string
    required:
    - driver_id
    type: object
  service.CancelShipmentInput:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  service.CreateShipmentInput:
    properties:
      cargo_description:
        type: string
      cargo_value:
        type: number
      cargo_weight:
        type: number
      customer_name:
        type: string
      customer_phone:
        type: string
      destination_address:
        type: string
      driver_id:
        type: string
      origin_address:
        type: string
      special_instructions:
        type: string
      vehicle_id:
        type: string
    required:
    - customer_name
    - customer_phone
    - destination_address
    - origin_address
    type: object
...
  service.UpdateShipmentInput:
    properties:
      cargo_description:
        type: string
      cargo_value:
        type: number
      cargo_weight:
        type: number
      customer_name:
        type: string
      customer_phone:
        type: string
      destination_address:
        type: string
      driver_id:
        type: string
      estimated_delivery:
        type: string
      origin_address:
        type: string
      special_instructions:
        type: string
      vehicle_id:
        type: string
    type: object
  service.UpdateStatusInput:
    properties:
      status:
        enum:
        - pending
        - assigned
        - in_transit
        - delayed
        - arrived
        - delivered
        - issue
        - cancelled
        type: string
      status_note:
        type: string
      status_reason:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@logistics.et
    name: API Support
  description: API for managing logistics operations including shipments, drivers,
    vehicles, and tracking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ethiopian Logistics Tracking Platform API
  version: "1.0"
paths:
...
  /my-shipments:
    get:
      description: Get active shipments for the logged-in driver
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Shipment'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List driver's shipments
      tags:
      - shipments
...
  /shipments:
    get:
      description: Get paginated list of shipments
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List shipments
      tags:
      - shipments
    post:
      consumes:
      - application/json
      description: Create a new shipment
      parameters:
      - description: Shipment data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.CreateShipmentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a shipment
      tags:
      - shipments
  /shipments/{id}:
    get:
      description: Get shipment details by ID
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get a shipment
      tags:
      - shipments
    put:
      consumes:
      - application/json
      description: Update shipment details
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Shipment data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.UpdateShipmentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a shipment
      tags:
      - shipments
  /shipments/{id}/assign:
    put:
      consumes:
      - application/json
      description: Assign a driver to a shipment
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.AssignDriverInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Assign driver to shipment
      tags:
      - shipments
  /shipments/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a shipment with a reason
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancellation reason
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.CancelShipmentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Cancel a shipment
      tags:
      - shipments
  /shipments/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a shipment
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Status data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update shipment status
      tags:
      - shipments
  /shipments/search:
    get:
      description: Search shipments with various filters
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by driver ID
        in: query
        name: driver_id
        type: string
      - description: Filter by origin address
        in: query
        name: origin
        type: string
      - description: Filter by destination address
        in: query
        name: destination
        type: string
      - description: Search query (tracking number, customer name, phone)
        in: query
        name: q
        type: string
      - description: From date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: To date (RFC3339)
        in: query
        name: date_to
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Search shipments
      tags:
      - shipments
...
  /track/{tracking_number}:
    get:
      description: Get shipment details by tracking number (public endpoint)
      parameters:
      - description: Tracking number
        in: path
        name: tracking_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shipment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Track a shipment
      tags:
      - tracking
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"